Style update - uint vs uint256

+++ smart-contracts/contracts/PublicLock.sol
@@ -97,7 +97,7 @@ contract PublicLock is ILockCore, ERC165, IERC721, IERC721Receiver, Ownable {
 
   // Ensure that the caller owns the key
   modifier onlyKeyOwner(
-    uint256 _tokenId
+    uint _tokenId
   ) {
     require(
       address(_tokenId) == msg.sender
@@ -109,7 +109,7 @@ contract PublicLock is ILockCore, ERC165, IERC721, IERC721Receiver, Ownable {
   // or that the caller has been approved
   // for ownership of that key
   modifier onlyKeyOwnerOrApproved(
-    uint256 _tokenId
+    uint _tokenId
   ) {
     require(
       address(_tokenId) == msg.sender
@@ -180,7 +180,7 @@ contract PublicLock is ILockCore, ERC165, IERC721, IERC721Receiver, Ownable {
   function transferFrom(
     address _from,
     address _recipient,
-    uint256 _tokenId
+    uint _tokenId
   )
     external
     payable
@@ -231,7 +231,7 @@ contract PublicLock is ILockCore, ERC165, IERC721, IERC721Receiver, Ownable {
     external
     onlyOwner
   {
-    uint256 balance = address(this).balance;
+    uint balance = address(this).balance;
     require(balance > 0, "Not enough funds");
     address owner = Ownable.owner();
     owner.transfer(balance);
@@ -244,7 +244,7 @@ contract PublicLock is ILockCore, ERC165, IERC721, IERC721Receiver, Ownable {
    */
   function approve(
     address _approved,
-    uint256 _tokenId
+    uint _tokenId
   )
     external
     payable
@@ -265,7 +265,7 @@ contract PublicLock is ILockCore, ERC165, IERC721, IERC721Receiver, Ownable {
     external
     onlyOwner
   {
-    uint256 oldKeyPrice = keyPrice;
+    uint oldKeyPrice = keyPrice;
     keyPrice = _keyPrice;
     emit PriceChanged(oldKeyPrice, keyPrice);
   }
@@ -279,7 +279,7 @@ contract PublicLock is ILockCore, ERC165, IERC721, IERC721Receiver, Ownable {
   )
     external
     view
-    returns (uint256)
+    returns (uint)
   {
     require(_owner != address(0), "Invalid address");
     return keyByOwner[_owner].expirationTimestamp > 0 ? 1 : 0;
@@ -290,7 +290,7 @@ contract PublicLock is ILockCore, ERC165, IERC721, IERC721Receiver, Ownable {
    * @return The address of the owner of the NFT, if applicable
   */
   function ownerOf(
-    uint256 _tokenId
+    uint _tokenId
   )
     external
     view
@@ -305,7 +305,7 @@ contract PublicLock is ILockCore, ERC165, IERC721, IERC721Receiver, Ownable {
    * Will return the approved recipient for a key, if any.
    */
   function getApproved(
-    uint256 _tokenId
+    uint _tokenId
   )
     external
     view
@@ -352,7 +352,7 @@ contract PublicLock is ILockCore, ERC165, IERC721, IERC721Receiver, Ownable {
     uint pageIndex = 0;
 
     // Build the requested set of owners into a new temporary array:
-    for (uint256 i = _startIndex; i < endOfPageIndex; i++) {
+    for (uint i = _startIndex; i < endOfPageIndex; i++) {
       ownersByPage[pageIndex] = owners[i];
       pageIndex++;
     }
@@ -415,12 +415,12 @@ contract PublicLock is ILockCore, ERC165, IERC721, IERC721Receiver, Ownable {
    * @param from The address which previously owned the token
    * @param tokenId The NFT identifier which is being transferred
    * @param data Additional data with no specified format
-   * @return `bytes4(keccak256("onERC721Received(address,address,uint256,bytes)"))`
+   * @return `bytes4(keccak256("onERC721Received(address,address,uint,bytes)"))`
    */
   function onERC721Received(
     address operator, // solhint-disable-line no-unused-vars
     address from, // solhint-disable-line no-unused-vars
-    uint256 tokenId, // solhint-disable-line no-unused-vars
+    uint tokenId, // solhint-disable-line no-unused-vars
     bytes data // solhint-disable-line no-unused-vars
   )
     public
@@ -491,7 +491,7 @@ contract PublicLock is ILockCore, ERC165, IERC721, IERC721Receiver, Ownable {
     emit Transfer(
       address(0), // This is a creation.
       _recipient,
-      uint256(_recipient) // Note: since each user can own a single token, we use the current
+      uint(_recipient) // Note: since each user can own a single token, we use the current
       // owner (new!) for the token id
     );
   }
@@ -502,7 +502,7 @@ contract PublicLock is ILockCore, ERC165, IERC721, IERC721Receiver, Ownable {
    * actually exists.
    */
   function _getApproved(
-    uint256 _tokenId
+    uint _tokenId
   )
     internal
     view
